refactor: reorganize api/auth/security into 2-depth subpackages by role
Split the flat security/ and jwt/ packages into 2-depth subpackages
grouped by Spring Security concept, so the role of each class is
obvious from its path instead of its filename.

security/ subpackage layout:
- config/       SecurityConfig, SecurityCorsProperties
- principal/    UserPrincipal
- annotation/   CurrentUser (already there from #157)
- filter/       JwtAuthenticationFilter
- userdetails/  CustomUserDetailsService
- handler/      JwtAccessDeniedHandler, JwtAuthenticationEntryPoint
- credential/   AuthenticationManagerCredentialAuthenticator

jwt/ now holds only pure JWT tech (no Spring Security types):
JwtProperties, JwtTokenUtil, JwtTokenProvider, ParsedAccessToken.

SecurityConfig moves from the slice root into security/config/. The
original reason for hoisting it to the root (avoiding a security -> jwt
-> security cycle) is gone now that JwtAuthenticationFilter lives
inside security/filter/ — the dependency graph is strictly
security/{config,filter,userdetails,handler,...} -> jwt -> security/principal,
with no back-edges into security/config/.

UserInfoService stays in jwt/ on purpose in this PR. It is slated for
deletion in a follow-up: its two public methods solve unrelated
problems (normal-flow actor lookup vs. auth-failure actor forensics)
and will be split into HandlerActorResolver (under security/handler/)
and CurrentRequestActorResolver (under api/common/logging/) in the
next PR. Moving it here would be throwaway work.
