+++ eventitta-api/src/main/java/com/eventitta/api/auth/security/config/SecurityConfig.java
@@ -1,11 +1,11 @@
-package com.eventitta.api.auth;
+package com.eventitta.api.auth.security.config;
 
 import com.eventitta.api.auth.jwt.JwtTokenProvider;
-import com.eventitta.api.auth.jwt.JwtAccessDeniedHandler;
-import com.eventitta.api.auth.jwt.JwtAuthenticationFilter;
-import com.eventitta.api.auth.jwt.CustomUserDetailsService;
-import com.eventitta.api.auth.jwt.JwtAuthenticationEntryPoint;
-import com.eventitta.api.auth.security.SecurityCorsProperties;
+import com.eventitta.api.auth.security.handler.JwtAccessDeniedHandler;
+import com.eventitta.api.auth.security.filter.JwtAuthenticationFilter;
+import com.eventitta.api.auth.security.userdetails.CustomUserDetailsService;
+import com.eventitta.api.auth.security.handler.JwtAuthenticationEntryPoint;
+import com.eventitta.api.auth.security.config.SecurityCorsProperties;
 import com.eventitta.domain.auth.repository.RefreshTokenRepository;
 import com.eventitta.domain.user.api.internal.facade.UserInternalFacade;
 import lombok.RequiredArgsConstructor;
