+++ eventitta-api/src/main/java/com/eventitta/api/auth/security/handler/JwtAccessDeniedHandler.java
@@ -1,7 +1,7 @@
-package com.eventitta.api.auth.jwt;
+package com.eventitta.api.auth.security.handler;
 
-import com.eventitta.domain.auth.exception.AuthErrorCode;
 import com.eventitta.api.auth.jwt.UserInfoService;
+import com.eventitta.domain.common.exception.CommonErrorCode;
 import com.eventitta.api.common.response.ApiErrorResponse;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import jakarta.servlet.http.HttpServletRequest;
@@ -10,8 +10,8 @@ import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.MediaType;
-import org.springframework.security.core.AuthenticationException;
-import org.springframework.security.web.AuthenticationEntryPoint;
+import org.springframework.security.access.AccessDeniedException;
+import org.springframework.security.web.access.AccessDeniedHandler;
 import org.springframework.stereotype.Component;
 
 import java.io.IOException;
@@ -19,44 +19,40 @@ import java.io.IOException;
 @Component
 @RequiredArgsConstructor
 @Slf4j
-public class JwtAuthenticationEntryPoint implements AuthenticationEntryPoint {
+public class JwtAccessDeniedHandler implements AccessDeniedHandler {
 
     private final ObjectMapper objectMapper;
     private final UserInfoService userInfoService;
 
     @Override
-    public void commence(HttpServletRequest request, HttpServletResponse response, AuthenticationException authException) throws IOException {
+    public void handle(
+        HttpServletRequest request,
+        HttpServletResponse response,
+        AccessDeniedException accessDeniedException
+    ) throws IOException {
         log.warn(
-            "[Exception] Authentication failure errorCode={}, status={}, path={}, userInfo={}, exceptionType={}, message={}",
-            AuthErrorCode.ACCESS_TOKEN_INVALID.name(),
-            HttpStatus.UNAUTHORIZED.value(),
+            "[Exception] Access denied errorCode={}, status={}, path={}, userInfo={}, exceptionType={}, message={}",
+            CommonErrorCode.FORBIDDEN.name(),
+            HttpStatus.FORBIDDEN.value(),
             request.getRequestURI(),
             resolveUserInfoSafely(request),
-            authException.getClass().getName(),
-            authException.getMessage()
+            accessDeniedException.getClass().getName(),
+            accessDeniedException.getMessage()
         );
         writeErrorResponse(request, response);
     }
 
     private void writeErrorResponse(HttpServletRequest request, HttpServletResponse response) throws IOException {
-        ApiErrorResponse body = createErrorResponseBody(request);
-        setResponseHeaders(response);
-        response.getWriter().write(objectMapper.writeValueAsString(body));
-    }
-
-    private ApiErrorResponse createErrorResponseBody(HttpServletRequest request) {
-        return ApiErrorResponse.of(
-            AuthErrorCode.ACCESS_TOKEN_INVALID.name(),
-            AuthErrorCode.ACCESS_TOKEN_INVALID.defaultMessage(),
-            HttpStatus.UNAUTHORIZED.value(),
+        ApiErrorResponse body = ApiErrorResponse.of(
+            CommonErrorCode.FORBIDDEN.name(),
+            CommonErrorCode.FORBIDDEN.defaultMessage(),
+            HttpStatus.FORBIDDEN.value(),
             request.getRequestURI()
         );
-    }
-
-    private void setResponseHeaders(HttpServletResponse response) {
-        response.setStatus(HttpStatus.UNAUTHORIZED.value());
+        response.setStatus(HttpStatus.FORBIDDEN.value());
         response.setContentType(MediaType.APPLICATION_JSON_VALUE);
         response.setCharacterEncoding("UTF-8");
+        response.getWriter().write(objectMapper.writeValueAsString(body));
     }
 
     private String resolveUserInfoSafely(HttpServletRequest request) {
