+++ eventitta-api/src/main/java/com/eventitta/api/post/controller/PostController.java
@@ -1,7 +1,7 @@
 package com.eventitta.api.post.controller;
 
 import com.eventitta.api.auth.security.annotation.CurrentUser;
-import com.eventitta.api.auth.security.UserPrincipal;
+import com.eventitta.api.auth.security.principal.UserPrincipal;
 import com.eventitta.domain.common.response.PageResponse;
 import com.eventitta.domain.post.dto.PostFilter;
 import com.eventitta.domain.post.dto.request.CreatePostRequest;
