+++ eventitta-api/src/main/java/com/eventitta/api/user/controller/UserController.java
@@ -2,7 +2,7 @@ package com.eventitta.api.user.controller;
 
 import com.eventitta.api.auth.security.annotation.CurrentUser;
 import com.eventitta.api.common.response.ApiErrorResponse;
-import com.eventitta.api.auth.security.UserPrincipal;
+import com.eventitta.api.auth.security.principal.UserPrincipal;
 import com.eventitta.api.auth.controller.request.SocialLoginRequest;
 import com.eventitta.api.auth.AuthConstants;
 import com.eventitta.api.auth.controller.AuthMapper;
